mcp-unstructured/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-unstructured"
version = "0.7.5"
description = "Lightweight MCP-style Unstructured document parser with optional hosted VLM mode"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy==1.26.4",
  "unstructured",
  "pdfminer.six",
  "requests",
  "python-docx",
  "python-pptx",
  "openpyxl",
  "pypdf",
  "pdf2image",
  "pytesseract",
  "unstructured-pytesseract",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
